Onboarding: split wow-funnel site creation out of the funnel model
CI caught this: site-spec's test suite stopped loading with "localize is not a
function". Importing the funnel helpers into site-spec dragged in a chain it had
never pulled before — wow-funnel.ts imports the '@automattic/onboarding' and
'@automattic/data-stores' barrels, which reach through cart, domain-search and
the component library to foldable-card, which needs i18n-calypso.

Both barrels exist for one thing: createSite. So separate the two halves along
the seam the funnel already has — creating the site, versus deciding where the
customer goes once it is built. wow-funnel-site.ts owns startWowFunnelSite,
forgetWowFunnelRun and the in-flight cache, and is imported only by the two
places that create sites. wow-funnel.ts keeps the model — slug, args, config,
destination, readiness, hand-off, session memory — and now imports nothing
heavier than the sibling utils it already used.

The hand-off side needed none of that weight; it was only ever inherited.

Narrowing the data-stores import to the site/types subpath is kept as well: it
is where Visibility actually lives, and the barrel was never warranted for one
enum.

Verified with the whole stepper suite rather than the handful of files I had
been running: 124 suites, 889 tests, all passing.
